First step refactoring before migrating to compositor custom visual rendering
